requirements.txt
pytube
assemblyai
srt_equalizer
python-dotenv
termcolor
revChatGPT
requests